+++ packages/cli/lib/project-generator.js
@@ -6,6 +6,7 @@
 'use strict';
 const BaseGenerator = require('./base-generator');
 const utils = require('./utils');
+const chalk = require('chalk');
 
 module.exports = class ProjectGenerator extends BaseGenerator {
   // Note: arguments and options should be defined in the constructor.
@@ -15,11 +16,24 @@ module.exports = class ProjectGenerator extends BaseGenerator {
     // This list gets shown to users to let them select the appropriate
     // build settings for their project.
     this.buildOptions = [
-      'tslint',
-      'prettier',
-      'mocha',
-      'loopbackBuild',
-      'vscode',
+      {
+        name: 'tslint',
+        description: 'add a linter with pre-configured lint rules',
+      },
+      {
+        name: 'prettier',
+        description:
+          'add new npm scripts to facilitate consistent code formatting',
+      },
+      {
+        name: 'mocha',
+        description: 'install mocha to assist with running tests',
+      },
+      {
+        name: 'loopbackBuild',
+        description: 'use @loopback/build helpers (e.g. lb-tslint)',
+      },
+      {name: 'vscode', description: 'add VSCode config files'},
     ];
   }
 
@@ -101,6 +115,9 @@ module.exports = class ProjectGenerator extends BaseGenerator {
       this.buildOptions,
     );
     this.projectOptions.forEach(n => {
+      if (typeof n === 'object') {
+        n = n.name;
+      }
       if (this.options[n]) {
         this.projectInfo[n] = this.options[n];
       }
@@ -158,10 +175,11 @@ module.exports = class ProjectGenerator extends BaseGenerator {
     if (this.shouldExit()) return false;
     const choices = [];
     this.buildOptions.forEach(f => {
-      if (!this.options[f]) {
+      if (!this.options[f.name]) {
         choices.push({
-          name: 'Enable ' + f,
-          key: f,
+          name: `Enable ${f.name}: ${chalk.gray(f.description)}`,
+          key: f.name,
+          short: `Enable ${f.name}`,
           checked: true,
         });
       }
@@ -177,11 +195,12 @@ module.exports = class ProjectGenerator extends BaseGenerator {
       },
     ];
     return this.prompt(prompts).then(props => {
-      const settings = props.settings || choices.map(c => c.name);
+      const settings = props.settings || choices.map(c => c.short);
       const features = choices.map(c => {
         return {
           key: c.key,
-          value: settings.indexOf(c.name) !== -1,
+          value:
+            settings.indexOf(c.name) !== -1 || settings.indexOf(c.short) !== -1,
         };
       });
       features.forEach(f => (this.projectInfo[f.key] = f.value));